Drop Timestamp type, use plain int64
The API returns Unix seconds and consumers convert anyway, so the
distinct wrapper was ceremony without payoff.

+++ src/openai/chat.nim
@@ -219,7 +219,7 @@ proc modelOf*(x: var ChatCreateResult): var string {.inline.} =
   result = x.model
 
 proc createdAt*(x: ChatCreateResult): int64 {.inline.} =
-  result = int64(x.created)
+  result = x.created
 
 proc promptTokens*(x: ChatCreateResult): int {.inline.} =
   result = x.usage.prompt_tokens
